requirements.txt
streamlit
openai
numpy
tenacity
python-dotenv
pathlib
pandas
plotly
kaleido